+++ packages/client/src/utils/log.ts
@@ -9,41 +9,6 @@ Debug.formatters.n = (v: any) => {
     if (v == null || Number.isNaN(v)) { return String(v) }
     return Debug.humanize(v)
 }
-
-export const DEFAULT_INSPECT_OPTS = {
-    maxStringLength: 256
-}
-
-declare const process: any
-
-const D: any = Debug
-
-// Override default formatters for environments with defined `Debug.inspectOpts` (only node atm).
-if ('inspectOpts' in D) {
-    // monkeypatch default log function to use current `inspectOpts`.  This
-    // ensures values logged without placeholders e.g. %o, %O will have the
-    // same inspect options applied. Without this only values with a
-    // placeholder will use the `inspectOpts` config.
-    // e.g.
-    // `debug('msg', obj)` should use same `inspectOpts` as `debug('msg %O', msg)`
-    D.log = function log(...args: any[]) {
-        const { inspectOpts, useColors } = this
-
-        if (inspectOpts.colors === undefined) {
-            inspectOpts.colors = useColors // need this to get colours when no placeholder
-        }
-
-        return process.stderr.write(formatWithOptions({
-            ...inspectOpts,
-        }, ...args) + '\n')
-    }
-
-    // mutate inspectOpts rather than replace, otherwise changes are lost
-    Object.assign(D.inspectOpts, {
-        ...DEFAULT_INSPECT_OPTS,
-    })
-}
-
 const streamrDebug = Debug('Streamr')
 
 const StreamrDebug = Object.assign(streamrDebug.extend.bind(streamrDebug), {
@@ -59,27 +24,18 @@ export {
 }
 
 // eslint-disable-next-line @typescript-eslint/explicit-module-boundary-types
-export function inspect(value: any, inspectOptions: Parameters<typeof util.inspect>[1] = {}): string {
-    return util.inspect(value, {
-        ...DEFAULT_INSPECT_OPTS,
-        ...inspectOptions,
-    })
-}
-
-// eslint-disable-next-line @typescript-eslint/explicit-module-boundary-types
-export function formatWithOptions(inspectOptions: Parameters<typeof util.formatWithOptions>[0], msgFormat?: any, ...param: any[]): string {
+export function formatWithOptions(msgFormat?: any, ...param: any[]): string {
     if (typeof util.formatWithOptions !== 'function') {
         // util.formatWithOptions is not browserified, use util.format instead
         return util.format(msgFormat, ...param)
     }
 
     return util.formatWithOptions({
-        ...DEFAULT_INSPECT_OPTS,
-        ...inspectOptions,
+        maxStringLength: 256,
     }, msgFormat, ...param)
 }
 
 // eslint-disable-next-line @typescript-eslint/explicit-module-boundary-types
 export function format(msgFormat?: any, ...param: any[]): string {
-    return formatWithOptions(DEFAULT_INSPECT_OPTS, msgFormat, ...param)
+    return formatWithOptions(msgFormat, ...param)
 }