+++ packages/client/test/integration/client-destroy.test.ts
@@ -1,20 +1,42 @@
 import 'reflect-metadata'
+import StreamrClient from '../../src'
+import { Stream } from '../../src/Stream'
 import { collect } from '../../src/utils/iterators'
 import { FakeEnvironment } from '../test-utils/fake/FakeEnvironment'
 import { createTestStream } from '../test-utils/utils'
 
 describe('client destroy', () => {
 
-    it('subscribe pipeline ends', async () => {
+    let client: StreamrClient
+    let stream: Stream
+
+    beforeEach(async () => {
         const environment = new FakeEnvironment()
-        const subscriber = environment.createClient()
-        const stream = await createTestStream(subscriber, module)
-        const sub = await subscriber.subscribe(stream.id)        
+        client = environment.createClient()
+        stream = await createTestStream(client, module)
+    })
+
+    it('ongoing subscribe pipeline ends', async () => {
+        const sub = await client.subscribe(stream.id)
         const onError: any = jest.fn()
         sub.on('error', onError)
         const outputPromise = collect(sub)
-        await subscriber.destroy()
+        await client.destroy()
         expect(onError).toBeCalledTimes(0)
         expect(await outputPromise).toEqual([])
+    })
+
+    it('unable to subscribe after destroy called', async () => {
+        await client.destroy()
+        await expect(async () => {
+            await client.subscribe(stream.id)
+        }).rejects.toThrow('Client is destroyed')
+    })
+
+    it('unable to publish after destroy called', async () => {
+        await client.destroy()
+        await expect(async () => {
+            await client.publish(stream.id, {})
+        }).rejects.toThrow(/Failed to publish.*Client is destroyed/)
     })
 })