+++ packages/client/test/test-utils/fake/FakeStorageNode.ts
@@ -56,7 +56,6 @@ export class FakeStorageNode extends FakeNetworkNode {
             metadata: {},
             permissions: storageNodeAssignmentStreamPermissions
         })
-    
     }
 
     async addAssignment(streamId: StreamID): Promise<void> {