+++ packages/client/CHANGELOG.md
@@ -50,6 +50,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - use methods `updateEncryptionKey` and `addEncryptionKey` instead
 - Remove client configuration option `verifySignatures`
 - Remove client configuration option `client.network.name`
+- Remove client configuration option `client.debug`
 - Remove (non-functional) client configuration options `autoConnect`, `autoDisconnect`, and `maxRetries`
 
 ### Fixed